+++ backend/node-http/index.js
@@ -1,0 +1,28 @@
+const http = require("http");
+
+const hostname = "localhost";
+const port = 3000;
+
+const server = http.createServer((req, res) => {
+  console.log(req.headers);
+
+  res.statusCode = 200;
+
+  res.setHeader("Content-Type", "text/html");
+  res.end(
+    `
+    <!DOCTYPE html>
+    <html>
+      <header>
+      </header>
+
+      <body>
+        <h1>Hello my gorgeous friend~</h1>
+      </body>
+    </html>`
+  );
+});
+
+server.listen(port, hostname, () => {
+  console.log(`Server running at: http://${hostname}:${port}`);
+});
