+++ backend/package.json
@@ -1,17 +1,17 @@
 {
-  "name": "Backend",
+  "name": "backend",
   "packageManager": "yarn@3.3.1",
   "version": "1.0.0",
   "description": "HTTP Modules and learning common practices",
   "main": "index.js",
   "scripts": {
-    "start": "node index"
+    "start": "node index",
+    "dev": "nodemon ./node-http/index.js"
   },
   "repository": "github.com/sancg/Full_stack_learning",
   "author": "Santiago",
   "license": "MIT",
   "devDependencies": {
     "nodemon": "^2.0.20"
-
   }
 }
